Reorganize resource selection components
Use consistent naming for mobile and desktop resource menu
components, and bring to one location.
